+++ @vue-start/element-pro/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 # 0.2.31
 
-- perf: 使用 selection 模式实现单选方案；
+- fix: Table selection 赋值期间不相应操作；
+- perf: Table 使用 selection 模式实现单选方案；
 - feat: ProDrawer
 
 # 0.2.29